proyecto/requirements.txt
django-crispy-forms
pillow